refactor(app-saves-domain)!: move entities to distinct namespace
