+++ src/SimulationStorm.AppSaves.Persistence/AppSaveRepository.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using SimulationStorm.AppSaves.Entities;
 
 namespace SimulationStorm.AppSaves.Persistence;
