+++ src/SimulationStorm.AppSaves.Presentation/ViewModels/AppSaveManagerDialogViewModel.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using DynamicData.Binding;
+using SimulationStorm.AppSaves.Entities;
 using SimulationStorm.AppSaves.Presentation.Models;
 using SimulationStorm.Dialogs.Presentation;
 using SimulationStorm.Notifications.Presentation;
