+++ src/SimulationStorm.AppSaves/AppSaveManager.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using DotNext.Collections.Generic;
 using DynamicData;
+using SimulationStorm.AppSaves.Entities;
 using SimulationStorm.AppSaves.Operations;
 
 namespace SimulationStorm.AppSaves;
