+++ src/SimulationStorm.AppSaves/IAppSaveRepository.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using SimulationStorm.AppSaves.Entities;
 
 namespace SimulationStorm.AppSaves;
